DataPipeline/requirements.txt
# TradeSmartAI 주식 데이터 수집 파이프라인 - Python 의존성

# 데이터베이스 관련
SQLAlchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0

# 데이터 처리
pandas>=1.5.0
numpy>=1.24.0

# 웹 요청 및 API
requests>=2.28.0

# 시간대 처리
pytz>=2022.1

# Apache Airflow
apache-airflow[postgres]>=2.6.0,<3.0.0